feat: 엔티티 클래스 - 리스트 자료형 정리
