+++ src/main/java/com/moodeat/domain/ManualStep.java
@@ -1,0 +1,7 @@
+package com.moodeat.domain;
+
+public class ManualStep {
+	private int order;
+	private String photo;
+	private String description;
+}
