+++ src/main/java/com/moodeat/domain/Message.java
@@ -1,0 +1,9 @@
+package com.moodeat.domain;
+
+import com.moodeat.domain.enums.MessageRole;
+
+public class Message {
+	private MessageRole role;
+	private String content;
+}
+
