+++ src/main/java/com/moodeat/domain/enums/MessageRole.java
@@ -1,0 +1,5 @@
+package com.moodeat.domain.enums;
+
+public enum MessageRole {
+	SYSTEM, ASSISTANT, USER
+}
