+++ src/main/java/com/moodeat/domain/enums/Mood.java
@@ -1,5 +1,14 @@
 package com.moodeat.domain.enums;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
 public enum Mood {
-	HAPPY, SAD, ANGRY
+	HAPPY("기쁨"),
+	ANGRY("화남"),
+	SAD("슬픔");
+
+	private final String mood;
 }
