--- src/main/java/com/moodeat/dto/ManualData.java
@@ -1,7 +1,0 @@
-package com.moodeat.dto;
-
-public class ManualData {
-	private int step;
-	private String photo;
-	private String description;
-}
