--- src/main/java/com/moodeat/dto/Message.java
@@ -1,6 +1,0 @@
-package com.moodeat.dto;
-
-public class Message {
-	private MessageType type;
-	private String content;
-}
