--- src/main/java/com/moodeat/dto/MessageType.java
@@ -1,5 +1,0 @@
-package com.moodeat.dto;
-
-public enum MessageType {
-	SYSTEM, ASSISTANT, USER
-}
